requirements.txt
fastapi==0.70.0
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]